files/default/requirements.txt
awscli
boto3
supervisor
requests
python-gnupg
jsonschema
jinja2
PyYAML
